Create generateAif.m
generateAif.m generates a simulated arterial input function (AIF). In the paper
describing the breast DRO, a publicly available AIF was used.  This generateAif
function was created to provide an AIF to the breast DRO in the event that the
AIF used in the breast DRO paper is unavailable to the user.  Using this AIF will
produce similar results to those obtained in the breast DRO paper.